Task refactoring to shutdown cleanly and remove test race conditions
