+++ nucleus/CMakeLists.txt
@@ -72,12 +72,13 @@ add_library(nucleus-obj OBJECT
         src/plugins/plugin_loader.cpp src/plugins/plugin_loader.hpp
         src/pubsub/local_topics.cpp src/pubsub/local_topics.hpp
         src/tasks/expire_time.hpp
-        src/tasks/sub_task.hpp
         src/tasks/task.cpp src/tasks/task.hpp
+        src/tasks/task_manager.cpp src/tasks/task_manager.hpp
+        src/tasks/task_threads.cpp src/tasks/task_threads.hpp
         src/util/commitable_file.cpp src/util/commitable_file.hpp
         src/util/nucleus_paths.cpp src/util/nucleus_paths.hpp
         src/util/permissions.cpp src/util/permissions.hpp
-)
+        )
 target_link_libraries(nucleus-obj
         PUBLIC gg_plugin_api
         PRIVATE Threads::Threads ${CMAKE_DL_LIBS} yaml-cpp)
