+++ nucleus/src/pubsub/local_topics.hpp
@@ -4,7 +4,7 @@
 #include "data/safe_handle.hpp"
 #include "data/shared_struct.hpp"
 #include "data/string_table.hpp"
-#include "tasks/sub_task.hpp"
+#include "tasks/task.hpp"
 #include <functional>
 #include <map>
 #include <string>
