+++ nucleus/src/tasks/task.hpp
@@ -1,21 +1,13 @@
 #pragma once
-#include "data/environment.hpp"
 #include "data/handle_table.hpp"
-#include "data/safe_handle.hpp"
-#include "data/shared_struct.hpp"
-#include "expire_time.hpp"
-#include "pubsub/local_topics.hpp"
-#include <condition_variable>
+#include "data/struct_model.hpp"
+#include "tasks/expire_time.hpp"
 #include <list>
-#include <mutex>
-#include <set>
-#include <thread>
-#include <vector>
 
 namespace tasks {
+    class SubTask;
     class TaskManager;
     class TaskThread;
-    class TaskPoolWorker;
     class Task;
     class TaskManager;
 
@@ -148,225 +140,25 @@ namespace tasks {
         bool terminatesWait();
 
         Status finalizeTask(const std::shared_ptr<data::StructModelBase> &data);
+        void requeueTask();
     };
 
-    class TaskThread : public util::RefObject<TaskThread> {
-        // mix-in representing either a worker thread or fixed thread
-
+    class SubTask : public util::RefObject<SubTask> {
     protected:
-        data::Environment &_environment;
-        std::weak_ptr<TaskManager> _pool;
-        std::list<std::shared_ptr<Task>> _tasks;
-        std::mutex _mutex;
-        std::condition_variable _wake;
-        bool _shutdown{false};
-        void bindThreadContext();
-
-        static TaskThread *getSetTaskThread(TaskThread *setValue, bool set) {
-            // NOLINTNEXTLINE(*-avoid-non-const-global-variables)
-            static thread_local TaskThread *_threadContext{nullptr};
-            TaskThread *current = _threadContext;
-            if(set) {
-                _threadContext = setValue;
-            }
-            return current;
-        }
+        std::shared_ptr<TaskThread> _threadAffinity;
 
     public:
-        explicit TaskThread(
-            data::Environment &environment, const std::shared_ptr<TaskManager> &pool
-        );
-        TaskThread(const TaskThread &) = delete;
-        TaskThread(TaskThread &&) = delete;
-        TaskThread &operator=(const TaskThread &) = delete;
-        TaskThread &operator=(TaskThread &&) = delete;
-        virtual ~TaskThread() = default;
-        void queueTask(const std::shared_ptr<Task> &task);
-        std::shared_ptr<Task> pickupAffinitizedTask();
-        std::shared_ptr<Task> pickupPoolTask();
-        std::shared_ptr<Task> pickupTask();
-        virtual void releaseFixedThread();
-
-        static std::shared_ptr<TaskThread> getThreadContext();
-
-        void shutdown() {
-            std::unique_lock guard(_mutex);
-            _shutdown = true;
-            _wake.notify_one();
-        }
-
-        void stall(const ExpireTime &end) {
-            std::unique_lock guard(_mutex);
-            if(_shutdown) {
-                return;
-            }
-            _wake.wait_until(guard, end.toTimePoint());
-        }
-
-        void waken() {
-            std::unique_lock guard(_mutex);
-            _wake.notify_one();
-        }
-
-        bool isShutdown() {
-            std::unique_lock guard(_mutex);
-            return _shutdown;
-        }
-
-        void taskStealing(const std::shared_ptr<Task> &blockedTask, const ExpireTime &end);
-        virtual ExpireTime taskStealingHook(ExpireTime time);
-
-        std::shared_ptr<Task> pickupTask(const std::shared_ptr<Task> &blockingTask);
-
-        std::shared_ptr<Task> pickupPoolTask(const std::shared_ptr<Task> &blockingTask);
+        SubTask() = default;
+        SubTask(const SubTask &) = delete;
+        SubTask(SubTask &&) = delete;
+        SubTask &operator=(const SubTask &) = delete;
+        SubTask &operator=(SubTask &&) = delete;
+        virtual ~SubTask() = default;
+        virtual std::shared_ptr<data::StructModelBase> runInThread(
+            const std::shared_ptr<Task> &task, const std::shared_ptr<data::StructModelBase> &dataIn
+        ) = 0;
+        void setAffinity(const std::shared_ptr<TaskThread> &affinity);
+        std::shared_ptr<TaskThread> getAffinity(const std::shared_ptr<TaskThread> &defaultThread);
     };
 
-    class TaskPoolWorker : public TaskThread {
-    private:
-        std::thread _thread;
-
-    public:
-        explicit TaskPoolWorker(
-            data::Environment &environment, const std::shared_ptr<TaskManager> &pool
-        );
-        void runner();
-    };
-
-    class FixedTaskThread : public TaskThread {
-    protected:
-        data::ObjectAnchor _defaultTask;
-        std::shared_ptr<FixedTaskThread> _protectThread;
-
-    public:
-        explicit FixedTaskThread(
-            data::Environment &environment, const std::shared_ptr<TaskManager> &pool
-        )
-            : TaskThread(environment, pool) {
-        }
-
-        // Call this on the native thread
-        void bindThreadContext(const data::ObjectAnchor &task);
-        void setDefaultTask(const data::ObjectAnchor &task);
-        data::ObjectAnchor getDefaultTask();
-        void protect();
-        void unprotect();
-        data::ObjectAnchor claimFixedThread();
-        void releaseFixedThread() override;
-    };
-
-    class FixedTimerTaskThread : public FixedTaskThread {
-    public:
-        explicit FixedTimerTaskThread(
-            data::Environment &environment, const std::shared_ptr<TaskManager> &pool
-        )
-            : FixedTaskThread(environment, pool) {
-        }
-
-        ExpireTime taskStealingHook(ExpireTime time) override;
-    };
-
-    class FixedTaskThreadScope {
-        std::shared_ptr<FixedTaskThread> _thread;
-
-    public:
-        FixedTaskThreadScope() = default;
-        FixedTaskThreadScope(const FixedTaskThreadScope &) = delete;
-        FixedTaskThreadScope &operator=(const FixedTaskThreadScope &) = delete;
-        FixedTaskThreadScope(FixedTaskThreadScope &&) noexcept = default;
-        FixedTaskThreadScope &operator=(FixedTaskThreadScope &&) = default;
-
-        ~FixedTaskThreadScope() {
-            release();
-        }
-
-        explicit FixedTaskThreadScope(const std::shared_ptr<FixedTaskThread> &thread)
-            : _thread(thread) {
-            if(thread) {
-                thread->claimFixedThread();
-            }
-        } // namespace tasks
-
-        void claim(const std::shared_ptr<FixedTaskThread> &thread) {
-            release();
-            _thread = thread;
-            if(thread) {
-                thread->claimFixedThread();
-            }
-        }
-
-        void release() {
-            if(_thread) {
-                _thread->releaseFixedThread();
-                _thread.reset();
-            }
-        }
-
-        [[nodiscard]] std::shared_ptr<FixedTaskThread> get() const {
-            return _thread;
-        }
-
-        [[nodiscard]] std::shared_ptr<Task> getTask() const {
-            return _thread->getDefaultTask().getObject<tasks::Task>();
-        }
-
-        explicit operator bool() {
-            return _thread.operator bool();
-        }
-
-        bool operator!() {
-            return !_thread;
-        }
-    };
-
-    inline void SubTask::setAffinity(const std::shared_ptr<TaskThread> &affinity) {
-        _threadAffinity = affinity;
-    }
-
-    inline std::shared_ptr<TaskThread> SubTask::getAffinity(
-        const std::shared_ptr<TaskThread> &defaultThread
-    ) {
-        std::shared_ptr<TaskThread> affinity = _threadAffinity;
-        if(affinity) {
-            return affinity;
-        } else {
-            return defaultThread;
-        }
-    }
-
-    class TaskManager : public data::TrackingScope {
-    private:
-        std::list<std::shared_ptr<TaskPoolWorker>> _busyWorkers; // assumes small
-                                                                 // pool, else
-                                                                 // std::set
-        std::list<std::shared_ptr<TaskPoolWorker>> _idleWorkers; // LIFO
-        std::list<std::shared_ptr<Task>> _backlog; // tasks with no thread affinity
-                                                   // (assumed async)
-        std::weak_ptr<TaskThread> _timerWorkerThread;
-        // _delayedTasks is using multimap as an insertable ordered list,
-        // TODO: is there a better std library for this?
-        std::multimap<ExpireTime, std::shared_ptr<Task>> _delayedTasks;
-        int _maxWorkers{5}; // TODO, from configuration
-
-        void scheduleFutureTaskAssumeLocked(
-            const ExpireTime &when, const std::shared_ptr<Task> &task
-        );
-        void descheduleFutureTaskAssumeLocked(
-            const ExpireTime &when, const std::shared_ptr<Task> &task
-        );
-        std::shared_ptr<Task> acquireTaskForWorker(TaskThread *worker);
-        std::shared_ptr<Task> acquireTaskWhenStealing(
-            TaskThread *worker, const std::shared_ptr<Task> &priorityTask
-        );
-        bool allocateNextWorker();
-        friend class Task;
-        friend class TaskThread;
-
-    public:
-        explicit TaskManager(data::Environment &environment) : data::TrackingScope{environment} {
-        }
-
-        data::ObjectAnchor createTask();
-        void queueTask(const std::shared_ptr<Task> &task);
-        ExpireTime pollNextDeferredTask(TaskThread *worker);
-    };
 } // namespace tasks
