+++ nucleus/tests/pubsub/pubsub_tests.cpp
@@ -1,9 +1,7 @@
 #include "data/globals.hpp"
 #include "pubsub/local_topics.hpp"
+#include "tasks/task_threads.hpp"
 #include <catch2/catch_all.hpp>
-#include <catch2/matchers/catch_matchers.hpp>
-#include <catch2/matchers/catch_matchers_string.hpp>
-#include <catch2/matchers/catch_matchers_vector.hpp>
 
 // NOLINTBEGIN
